Moved Alert related calls and types to seperate file
