+++ gvm/protocols/gmpv208/gmpv208.py
@@ -60,95 +60,6 @@ PROTOCOL_VERSION = (20, 8)
 
 logger = logging.getLogger(__name__)
 
-# put this into the Alert Entity
-def _check_event(
-    event: AlertEvent, condition: AlertCondition, method: AlertMethod
-):
-    if event == AlertEvent.TASK_RUN_STATUS_CHANGED:
-        if not condition:
-            raise RequiredArgument(
-                "condition is required for event {}".format(event.name)
-            )
-
-        if not method:
-            raise RequiredArgument(
-                "method is required for event {}".format(event.name)
-            )
-
-        if condition not in (
-            AlertCondition.ALWAYS,
-            AlertCondition.FILTER_COUNT_CHANGED,
-            AlertCondition.FILTER_COUNT_AT_LEAST,
-            AlertCondition.SEVERITY_AT_LEAST,
-            AlertCondition.SEVERITY_CHANGED,
-        ):
-            raise InvalidArgument(
-                "Invalid condition {} for event {}".format(
-                    condition.name, event.name
-                )
-            )
-    elif event in (
-        AlertEvent.NEW_SECINFO_ARRIVED,
-        AlertEvent.UPDATED_SECINFO_ARRIVED,
-    ):
-        if not condition:
-            raise RequiredArgument(
-                "condition is required for event {}".format(event.name)
-            )
-
-        if not method:
-            raise RequiredArgument(
-                "method is required for event {}".format(event.name)
-            )
-
-        if condition != AlertCondition.ALWAYS:
-            raise InvalidArgument(
-                "Invalid condition {} for event {}".format(
-                    condition.name, event.name
-                )
-            )
-        if method not in (
-            AlertMethod.SCP,
-            AlertMethod.SEND,
-            AlertMethod.SMB,
-            AlertMethod.SNMP,
-            AlertMethod.SYSLOG,
-            AlertMethod.EMAIL,
-        ):
-            raise InvalidArgument(
-                "Invalid method {} for event {}".format(method.name, event.name)
-            )
-    elif event in (
-        AlertEvent.TICKET_RECEIVED,
-        AlertEvent.OWNED_TICKET_CHANGED,
-        AlertEvent.ASSIGNED_TICKET_CHANGED,
-    ):
-        if not condition:
-            raise RequiredArgument(
-                "condition is required for event {}".format(event.name)
-            )
-
-        if not method:
-            raise RequiredArgument(
-                "method is required for event {}".format(event.name)
-            )
-        if condition != AlertCondition.ALWAYS:
-            raise InvalidArgument(
-                "Invalid condition {} for event {}".format(
-                    condition.name, event.name
-                )
-            )
-        if method not in (
-            AlertMethod.EMAIL,
-            AlertMethod.START_TASK,
-            AlertMethod.SYSLOG,
-        ):
-            raise InvalidArgument(
-                "Invalid method {} for event {}".format(method.name, event.name)
-            )
-    elif event is not None:
-        raise InvalidArgument('Invalid event "{}"'.format(event.name))
-
 
 class GmpV208Mixin(GvmProtocol):
     """Python interface for Greenbone Management Protocol
@@ -232,120 +143,6 @@ class GmpV208Mixin(GvmProtocol):
             self._authenticated = True
 
         return self._transform(response)
-
-    def create_alert(
-        self,
-        name: str,
-        condition: AlertCondition,
-        event: AlertEvent,
-        method: AlertMethod,
-        *,
-        method_data: Optional[dict] = None,
-        event_data: Optional[dict] = None,
-        condition_data: Optional[dict] = None,
-        filter_id: Optional[int] = None,
-        comment: Optional[str] = None,
-    ) -> Any:
-        """Create a new alert
-
-        Arguments:
-            name: Name of the new Alert
-            condition: The condition that must be satisfied for the alert
-                to occur; if the event is either 'Updated SecInfo arrived' or
-                'New SecInfo arrived', condition must be 'Always'. Otherwise,
-                condition can also be on of 'Severity at least', 'Filter count
-                changed' or 'Filter count at least'.
-            event: The event that must happen for the alert to occur, one
-                of 'Task run status changed', 'Updated SecInfo arrived' or 'New
-                SecInfo arrived'
-            method: The method by which the user is alerted, one of 'SCP',
-                'Send', 'SMB', 'SNMP', 'Syslog' or 'Email'; if the event is
-                neither 'Updated SecInfo arrived' nor 'New SecInfo arrived',
-                method can also be one of 'Start Task', 'HTTP Get', 'Sourcefire
-                Connector' or 'verinice Connector'.
-            condition_data: Data that defines the condition
-            event_data: Data that defines the event
-            method_data: Data that defines the method
-            filter_id: Filter to apply when executing alert
-            comment: Comment for the alert
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        if not name:
-            raise RequiredArgument(
-                function=self.create_alert.__name__, argument='name'
-            )
-
-        if not condition:
-            raise RequiredArgument(
-                function=self.create_alert.__name__, argument='condition'
-            )
-
-        if not event:
-            raise RequiredArgument(
-                function=self.create_alert.__name__, argument='event'
-            )
-
-        if not method:
-            raise RequiredArgument(
-                function=self.create_alert.__name__, argument='method'
-            )
-
-        if not isinstance(condition, AlertCondition):
-            raise InvalidArgumentType(
-                function=self.create_alert.__name__,
-                argument='condition',
-                arg_type=AlertCondition.__name__,
-            )
-
-        if not isinstance(event, AlertEvent):
-            raise InvalidArgumentType(
-                function=self.create_alert.__name__,
-                argument='even',
-                arg_type=AlertEvent.__name__,
-            )
-
-        if not isinstance(method, AlertMethod):
-            raise InvalidArgumentType(
-                function=self.create_alert.__name__,
-                argument='method',
-                arg_type=AlertMethod.__name__,
-            )
-
-        _check_event(event, condition, method)
-
-        cmd = XmlCommand("create_alert")
-        cmd.add_element("name", name)
-
-        conditions = cmd.add_element("condition", condition.value)
-
-        if condition_data is not None:
-            for key, value in condition_data.items():
-                _data = conditions.add_element("data", value)
-                _data.add_element("name", key)
-
-        events = cmd.add_element("event", event.value)
-
-        if event_data is not None:
-            for key, value in event_data.items():
-                _data = events.add_element("data", value)
-                _data.add_element("name", key)
-
-        methods = cmd.add_element("method", method.value)
-
-        if method_data is not None:
-            for key, value in method_data.items():
-                _data = methods.add_element("data", value)
-                _data.add_element("name", key)
-
-        if filter_id:
-            cmd.add_element("filter", attrs={"id": filter_id})
-
-        if comment:
-            cmd.add_element("comment", comment)
-
-        return self._send_xml_command(cmd)
 
     def create_audit(
         self,
@@ -908,114 +705,6 @@ class GmpV208Mixin(GvmProtocol):
 
         return self._send_xml_command(cmd)
 
-    def modify_alert(
-        self,
-        alert_id: str,
-        *,
-        name: Optional[str] = None,
-        comment: Optional[str] = None,
-        filter_id: Optional[str] = None,
-        event: Optional[AlertEvent] = None,
-        event_data: Optional[dict] = None,
-        condition: Optional[AlertCondition] = None,
-        condition_data: Optional[dict] = None,
-        method: Optional[AlertMethod] = None,
-        method_data: Optional[dict] = None,
-    ) -> Any:
-        """Modifies an existing alert.
-
-        Arguments:
-            alert_id: UUID of the alert to be modified.
-            name: Name of the Alert.
-            condition: The condition that must be satisfied for the alert to
-                occur. If the event is either 'Updated SecInfo
-                arrived' or 'New SecInfo arrived', condition must be 'Always'.
-                Otherwise, condition can also be on of 'Severity at least',
-                'Filter count changed' or 'Filter count at least'.
-            condition_data: Data that defines the condition
-            event: The event that must happen for the alert to occur, one of
-                'Task run status changed', 'Updated SecInfo arrived' or
-                'New SecInfo arrived'
-            event_data: Data that defines the event
-            method: The method by which the user is alerted, one of 'SCP',
-                'Send', 'SMB', 'SNMP', 'Syslog' or 'Email';
-                if the event is neither 'Updated SecInfo arrived' nor
-                'New SecInfo arrived', method can also be one of 'Start Task',
-                'HTTP Get', 'Sourcefire Connector' or 'verinice Connector'.
-            method_data: Data that defines the method
-            filter_id: Filter to apply when executing alert
-            comment: Comment for the alert
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-
-        if not alert_id:
-            raise RequiredArgument(
-                function=self.modify_alert.__name__, argument='alert_id'
-            )
-
-        cmd = XmlCommand("modify_alert")
-        cmd.set_attribute("alert_id", str(alert_id))
-
-        if name:
-            cmd.add_element("name", name)
-
-        if comment:
-            cmd.add_element("comment", comment)
-
-        if filter_id:
-            cmd.add_element("filter", attrs={"id": filter_id})
-
-        if condition:
-            if not isinstance(condition, AlertCondition):
-                raise InvalidArgumentType(
-                    function=self.modify_alert.__name__,
-                    argument='condition',
-                    arg_type=AlertCondition.__name__,
-                )
-
-            conditions = cmd.add_element("condition", condition.value)
-
-            if condition_data is not None:
-                for key, value in condition_data.items():
-                    _data = conditions.add_element("data", value)
-                    _data.add_element("name", key)
-
-        if method:
-            if not isinstance(method, AlertMethod):
-                raise InvalidArgumentType(
-                    function=self.modify_alert.__name__,
-                    argument='method',
-                    arg_type=AlertMethod.__name__,
-                )
-
-            methods = cmd.add_element("method", method.value)
-
-            if method_data is not None:
-                for key, value in method_data.items():
-                    _data = methods.add_element("data", value)
-                    _data.add_element("name", key)
-
-        if event:
-            if not isinstance(event, AlertEvent):
-                raise InvalidArgumentType(
-                    function=self.modify_alert.__name__,
-                    argument='event',
-                    arg_type=AlertEvent.__name__,
-                )
-
-            _check_event(event, condition, method)
-
-            events = cmd.add_element("event", event.value)
-
-            if event_data is not None:
-                for key, value in event_data.items():
-                    _data = events.add_element("data", value)
-                    _data.add_element("name", key)
-
-        return self._send_xml_command(cmd)
-
     def modify_audit(
         self,
         audit_id: str,
@@ -1383,24 +1072,6 @@ class GmpV208Mixin(GvmProtocol):
         if trust:
             cmd.add_element("trust", to_bool(trust))
 
-        return self._send_xml_command(cmd)
-
-    def clone_alert(self, alert_id: str) -> Any:
-        """Clone an existing alert
-
-        Arguments:
-            alert_id: UUID of an existing alert to clone from
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        if not alert_id:
-            raise RequiredArgument(
-                function=self.clone_alert.__name__, argument='alert_id'
-            )
-
-        cmd = XmlCommand("create_alert")
-        cmd.add_element("copy", alert_id)
         return self._send_xml_command(cmd)
 
     def clone_ticket(self, ticket_id: str) -> Any:
@@ -3360,26 +3031,6 @@ class GmpV208Mixin(GvmProtocol):
         cmd.add_element("copy", user_id)
         return self._send_xml_command(cmd)
 
-    def delete_alert(
-        self, alert_id: str, *, ultimate: Optional[bool] = False
-    ) -> Any:
-        """Deletes an existing alert
-
-        Arguments:
-            alert_id: UUID of the alert to be deleted.
-            ultimate: Whether to remove entirely, or to the trashcan.
-        """
-        if not alert_id:
-            raise RequiredArgument(
-                function=self.delete_alert.__name__, argument='alert_id'
-            )
-
-        cmd = XmlCommand("delete_alert")
-        cmd.set_attribute("alert_id", alert_id)
-        cmd.set_attribute("ultimate", to_bool(ultimate))
-
-        return self._send_xml_command(cmd)
-
     def delete_asset(
         self, *, asset_id: Optional[str] = None, report_id: Optional[str] = None
     ) -> Any:
@@ -3677,59 +3328,6 @@ class GmpV208Mixin(GvmProtocol):
             The response. See :py:meth:`send_command` for details.
         """
         return self._send_xml_command(XmlCommand("empty_trashcan"))
-
-    def get_alerts(
-        self,
-        *,
-        filter: Optional[str] = None,
-        filter_id: Optional[str] = None,
-        trash: Optional[bool] = None,
-        tasks: Optional[bool] = None,
-    ) -> Any:
-        """Request a list of alerts
-
-        Arguments:
-            filter: Filter term to use for the query
-            filter_id: UUID of an existing filter to use for the query
-            trash: True to request the alerts in the trashcan
-            tasks: Whether to include the tasks using the alerts
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        cmd = XmlCommand("get_alerts")
-
-        add_filter(cmd, filter, filter_id)
-
-        if trash is not None:
-            cmd.set_attribute("trash", to_bool(trash))
-
-        if tasks is not None:
-            cmd.set_attribute("tasks", to_bool(tasks))
-
-        return self._send_xml_command(cmd)
-
-    def get_alert(self, alert_id: str, *, tasks: Optional[bool] = None) -> Any:
-        """Request a single alert
-
-        Arguments:
-            alert_id: UUID of an existing alert
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        cmd = XmlCommand("get_alerts")
-
-        if not alert_id:
-            raise RequiredArgument(
-                function=self.get_alert.__name__, argument='alert_id'
-            )
-
-        cmd.set_attribute("alert_id", alert_id)
-
-        if tasks is not None:
-            cmd.set_attribute("tasks", to_bool(tasks))
-
-        return self._send_xml_command(cmd)
 
     def get_assets(
         self,
@@ -5280,85 +4878,6 @@ class GmpV208Mixin(GvmProtocol):
             The response. See :py:meth:`send_command` for details.
         """
         return self._send_xml_command(XmlCommand("sync_scap"))
-
-    def test_alert(self, alert_id: str) -> Any:
-        """Run an alert
-
-        Invoke a test run of an alert
-
-        Arguments:
-            alert_id: UUID of the alert to be tested
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        if not alert_id:
-            raise InvalidArgument("test_alert requires an alert_id argument")
-
-        cmd = XmlCommand("test_alert")
-        cmd.set_attribute("alert_id", alert_id)
-
-        return self._send_xml_command(cmd)
-
-    def trigger_alert(
-        self,
-        alert_id: str,
-        report_id: str,
-        *,
-        filter: Optional[str] = None,
-        filter_id: Optional[str] = None,
-        report_format_id: Optional[Union[str, ReportFormatType]] = None,
-        delta_report_id: Optional[str] = None,
-    ) -> Any:
-        """Run an alert by ignoring its event and conditions
-
-        The alert is triggered to run immediately with the provided filtered
-        report by ignoring the even and condition settings.
-
-        Arguments:
-            alert_id: UUID of the alert to be run
-            report_id: UUID of the report to be provided to the alert
-            filter: Filter term to use to filter results in the report
-            filter_id: UUID of filter to use to filter results in the report
-            report_format_id: UUID of report format to use
-                              or ReportFormatType (enum)
-            delta_report_id: UUID of an existing report to compare report to.
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        if not alert_id:
-            raise RequiredArgument(
-                function=self.trigger_alert.__name__,
-                argument='alert_id argument',
-            )
-
-        if not report_id:
-            raise RequiredArgument(
-                function=self.trigger_alert.__name__,
-                argument='report_id argument',
-            )
-
-        cmd = XmlCommand("get_reports")
-        cmd.set_attribute("report_id", report_id)
-        cmd.set_attribute("alert_id", alert_id)
-
-        if filter:
-            cmd.set_attribute("filter", filter)
-
-        if filter_id:
-            cmd.set_attribute("filt_id", filter_id)
-
-        if report_format_id:
-            if isinstance(report_format_id, ReportFormatType):
-                report_format_id = report_format_id.value
-
-            cmd.set_attribute("format_id", report_format_id)
-
-        if delta_report_id:
-            cmd.set_attribute("delta_report_id", delta_report_id)
-
-        return self._send_xml_command(cmd)
 
     def verify_report_format(
         self, report_format_id: Union[str, ReportFormatType]
